services/anomaly-service/requirements.txt
clickhouse-connect==0.8.18
python-dotenv==1.0.1
pydantic>=2.0.0